+++ src/wai/annotations/tf/compat/v1/__init__.py
@@ -1,0 +1,4 @@
+"""
+Package for Tensorflow V1 code compatibility.
+"""
+from ._ensure_eager_execution import ensure_eager_execution