+++ src/wai/annotations/tf/compat/v2/_ensure_eager_execution.py
@@ -1,0 +1,8 @@
+def ensure_eager_execution():
+    """
+    Ensures that Tensorflow is in eager-execution mode.
+    """
+    # Tensorflow V2 is eager by default
+    from tensorflow import executing_eagerly
+    if not executing_eagerly():
+        raise RuntimeError(f"Tensorflow V2 not in eager-execution mode")